+++ changelog.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
 
+- Page SaveAs, SaveTo and WriteTo use markdown for .md suffix in filename
+- Renamed MarkdownWriter to MarkdownEncoder
 - Renamed HtmlWriter to HtmlEncoder and WriteTo method to Encode
 - Render A tags in markdown
 - MakeTOC adds links for named headers for easier sharing